+++ changelog.md
@@ -5,6 +5,7 @@
 **Added:**
 - bpk-tokens:
   - iOS & Android border size tokens
+  - Android elevation tokens
 
 ## 2017-10-06 - React Native Icon Component
 - react-native-bpk-component-icon: 0.0.4 => 1.0.0